fix salary variable type

+++ src/entities/Employee.java
@@ -2,12 +2,12 @@ package entities;
 
 public class Employee {
     private String name;
-    private int salary;
+    private double salary;
     private int workHours; //haftalik calisma saati
     private int hireYear;
 
     //constructor
-    public Employee(String name, int salary, int workHours, int hireYear) {
+    public Employee(String name, double salary, int workHours, int hireYear) {
         this.name = name;
         this.salary = salary;
         this.workHours = workHours;
@@ -23,11 +23,11 @@ public class Employee {
         this.name = name;
     }
 
-    public int getSalary() {
+    public double getSalary() {
         return salary;
     }
 
-    public void setSalary(int salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
